% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-package.R
\docType{package}
\name{family-package}
\alias{family}
\alias{family-package}
\title{family: A Simpler Way to Find Related Directories}
\description{
Find a family of related directories without
    knowing their name.  Find children directories under a parent
    directory, the siblings of any one child, and the parent of all
    children.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/family}
  \item Report bugs at \url{https://github.com/maurolepore/family/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID}) [contractor]

}
\keyword{internal}
